refactor: useBackClose 를 전역 훅으로 이동
storeDetail 전용이 아니게 되어 features 밖으로 뺀다.
건물 층 오버레이도 같은 훅으로 뒤로가기를 받는다.
